refactor(revm): isolate signature gas accounting
